Renamed all nav components to correct typo
