+++ assets/constants/navigationConstants.js
